refactor: move tracer-related code into sub-modules
