+++ python/openinference-instrumentation/src/openinference/instrumentation/__init__.py
@@ -1,8 +1,8 @@
+from ._tracer_providers import TracerProvider
+from ._tracers import OITracer
 from .config import (
     REDACTED_VALUE,
-    OITracer,
     TraceConfig,
-    TracerProvider,
     get_input_attributes,
     get_output_attributes,
     get_span_kind_attributes,
