+++ python/openinference-instrumentation/src/openinference/instrumentation/_spans.py
@@ -1,0 +1,83 @@
+from typing import Any, Callable, Dict, Optional, Union, cast
+
+import wrapt  # type: ignore[import-untyped]
+from opentelemetry.trace import Span
+from opentelemetry.util.types import AttributeValue
+
+from .config import (
+    _IMPORTANT_ATTRIBUTES,
+    OPENINFERENCE_SPAN_KIND,
+    TOOL,
+    OpenInferenceMimeType,
+    TraceConfig,
+    get_input_attributes,
+    get_output_attributes,
+    get_tool_attributes,
+)
+
+
+class OpenInferenceSpan(wrapt.ObjectProxy):  # type: ignore[misc]
+    def __init__(self, wrapped: Span, config: TraceConfig) -> None:
+        super().__init__(wrapped)
+        self._self_config = config
+        self._self_important_attributes: Dict[str, AttributeValue] = {}
+
+    def set_attributes(self, attributes: Dict[str, AttributeValue]) -> None:
+        for k, v in attributes.items():
+            self.set_attribute(k, v)
+
+    def set_attribute(
+        self,
+        key: str,
+        value: Union[AttributeValue, Callable[[], AttributeValue]],
+    ) -> None:
+        masked_value = self._self_config.mask(key, value)
+        if masked_value is not None:
+            if key in _IMPORTANT_ATTRIBUTES:
+                self._self_important_attributes[key] = masked_value
+            else:
+                span = cast(Span, self.__wrapped__)
+                span.set_attribute(key, masked_value)
+
+    def end(self, end_time: Optional[int] = None) -> None:
+        span = cast(Span, self.__wrapped__)
+        for k, v in reversed(self._self_important_attributes.items()):
+            span.set_attribute(k, v)
+        span.end(end_time)
+
+    def set_input(
+        self,
+        value: Any,
+        *,
+        mime_type: Optional[OpenInferenceMimeType] = None,
+    ) -> None:
+        if OPENINFERENCE_SPAN_KIND not in self._self_important_attributes:
+            raise ValueError("Cannot set input attributes on a non-OpenInference span")
+        self.set_attributes(get_input_attributes(value, mime_type=mime_type))
+
+    def set_output(
+        self,
+        value: Any,
+        *,
+        mime_type: Optional[OpenInferenceMimeType] = None,
+    ) -> None:
+        if OPENINFERENCE_SPAN_KIND not in self._self_important_attributes:
+            raise ValueError("Cannot set output attributes on a non-OpenInference span")
+        self.set_attributes(get_output_attributes(value, mime_type=mime_type))
+
+    def set_tool(
+        self,
+        *,
+        name: str,
+        description: Optional[str] = None,
+        parameters: Union[str, Dict[str, Any]],
+    ) -> None:
+        if self._self_important_attributes.get(OPENINFERENCE_SPAN_KIND) != TOOL:
+            raise ValueError("Cannot set tool attributes on a non-tool span")
+        self.set_attributes(
+            get_tool_attributes(
+                name=name,
+                description=description,
+                parameters=parameters,
+            )
+        )
