+++ python/openinference-instrumentation/src/openinference/instrumentation/config.py
@@ -1,4 +1,3 @@
-import asyncio
 import collections.abc
 import inspect
 import json
@@ -13,7 +12,6 @@ from types import ModuleType, TracebackType
 from typing import (  #  type: ignore[attr-defined]
     TYPE_CHECKING,
     Any,
-    Awaitable,
     Callable,
     Dict,
     Iterator,
@@ -24,36 +22,25 @@ from typing import (  #  type: ignore[attr-defined]
     TypeVar,
     Union,
     _TypedDictMeta,
-    cast,
     get_args,
     get_origin,
 )
 
-import wrapt  # type: ignore[import-untyped]
 from opentelemetry.context import (
     _SUPPRESS_INSTRUMENTATION_KEY,
-    Context,
     attach,
     detach,
-    get_value,
     set_value,
 )
-from opentelemetry.sdk.trace import TracerProvider as OTelTracerProvider
-from opentelemetry.sdk.trace.id_generator import IdGenerator, RandomIdGenerator
+from opentelemetry.sdk.trace.id_generator import IdGenerator
 from opentelemetry.trace import (
-    INVALID_SPAN,
     INVALID_SPAN_ID,
     INVALID_TRACE_ID,
-    Link,
-    Span,
-    SpanKind,
     Status,
     StatusCode,
-    Tracer,
-    use_span,
 )
-from opentelemetry.util.types import Attributes, AttributeValue
-from typing_extensions import ParamSpec, TypeAlias, TypeGuard, _AnnotatedAlias, overload
+from opentelemetry.util.types import AttributeValue
+from typing_extensions import ParamSpec, TypeAlias, TypeGuard, _AnnotatedAlias
 
 from openinference.semconv.trace import (
     EmbeddingAttributes,
@@ -65,11 +52,14 @@ from openinference.semconv.trace import (
     SpanAttributes,
 )
 
-from .context_attributes import get_attributes_from_context
 from .logging import logger
 
 if TYPE_CHECKING:
     from _typeshed import DataclassInstance
+
+    from ._spans import OpenInferenceSpan
+    from ._tracers import OITracer
+
 pydantic: Optional[ModuleType]
 try:
     import pydantic  # try to import without adding a dependency
@@ -520,73 +510,6 @@ def get_tool_attributes(
     return attributes
 
 
-class OpenInferenceSpan(wrapt.ObjectProxy):  # type: ignore[misc]
-    def __init__(self, wrapped: Span, config: TraceConfig) -> None:
-        super().__init__(wrapped)
-        self._self_config = config
-        self._self_important_attributes: Dict[str, AttributeValue] = {}
-
-    def set_attributes(self, attributes: Dict[str, AttributeValue]) -> None:
-        for k, v in attributes.items():
-            self.set_attribute(k, v)
-
-    def set_attribute(
-        self,
-        key: str,
-        value: Union[AttributeValue, Callable[[], AttributeValue]],
-    ) -> None:
-        masked_value = self._self_config.mask(key, value)
-        if masked_value is not None:
-            if key in _IMPORTANT_ATTRIBUTES:
-                self._self_important_attributes[key] = masked_value
-            else:
-                span = cast(Span, self.__wrapped__)
-                span.set_attribute(key, masked_value)
-
-    def end(self, end_time: Optional[int] = None) -> None:
-        span = cast(Span, self.__wrapped__)
-        for k, v in reversed(self._self_important_attributes.items()):
-            span.set_attribute(k, v)
-        span.end(end_time)
-
-    def set_input(
-        self,
-        value: Any,
-        *,
-        mime_type: Optional[OpenInferenceMimeType] = None,
-    ) -> None:
-        if OPENINFERENCE_SPAN_KIND not in self._self_important_attributes:
-            raise ValueError("Cannot set input attributes on a non-OpenInference span")
-        self.set_attributes(get_input_attributes(value, mime_type=mime_type))
-
-    def set_output(
-        self,
-        value: Any,
-        *,
-        mime_type: Optional[OpenInferenceMimeType] = None,
-    ) -> None:
-        if OPENINFERENCE_SPAN_KIND not in self._self_important_attributes:
-            raise ValueError("Cannot set output attributes on a non-OpenInference span")
-        self.set_attributes(get_output_attributes(value, mime_type=mime_type))
-
-    def set_tool(
-        self,
-        *,
-        name: str,
-        description: Optional[str] = None,
-        parameters: Union[str, Dict[str, Any]],
-    ) -> None:
-        if self._self_important_attributes.get(OPENINFERENCE_SPAN_KIND) != TOOL:
-            raise ValueError("Cannot set tool attributes on a non-tool span")
-        self.set_attributes(
-            get_tool_attributes(
-                name=name,
-                description=description,
-                parameters=parameters,
-            )
-        )
-
-
 class _IdGenerator(IdGenerator):
     """
     An IdGenerator that uses a different source of randomness to
@@ -602,313 +525,6 @@ class _IdGenerator(IdGenerator):
         while (trace_id := randbits(128)) == INVALID_TRACE_ID:
             continue
         return trace_id
-
-
-class OITracer(wrapt.ObjectProxy):  # type: ignore[misc]
-    def __init__(self, wrapped: Tracer, config: TraceConfig) -> None:
-        super().__init__(wrapped)
-        self._self_config = config
-        self._self_id_generator = _IdGenerator()
-
-    @property
-    def id_generator(self) -> IdGenerator:
-        ans = getattr(self.__wrapped__, "id_generator", None)
-        if ans and ans.__class__ is RandomIdGenerator:
-            return self._self_id_generator
-        return cast(IdGenerator, ans)
-
-    @contextmanager
-    def start_as_current_span(
-        self,
-        name: str,
-        context: Optional[Context] = None,
-        kind: SpanKind = SpanKind.INTERNAL,
-        attributes: Attributes = None,
-        links: Optional["Sequence[Link]"] = (),
-        start_time: Optional[int] = None,
-        record_exception: bool = True,
-        set_status_on_exception: bool = True,
-        end_on_exit: bool = True,
-        *,
-        openinference_span_kind: Optional[OpenInferenceSpanKind] = None,
-    ) -> Iterator[OpenInferenceSpan]:
-        span = self.start_span(
-            name=name,
-            openinference_span_kind=openinference_span_kind,
-            context=context,
-            kind=kind,
-            attributes=attributes,
-            links=links,
-            start_time=start_time,
-            record_exception=record_exception,
-            set_status_on_exception=set_status_on_exception,
-        )
-        with use_span(
-            span,
-            end_on_exit=end_on_exit,
-            record_exception=record_exception,
-            set_status_on_exception=set_status_on_exception,
-        ) as current_span:
-            yield cast(OpenInferenceSpan, current_span)
-
-    def start_span(
-        self,
-        name: str,
-        context: Optional[Context] = None,
-        kind: SpanKind = SpanKind.INTERNAL,
-        attributes: Attributes = None,
-        links: Optional["Sequence[Link]"] = (),
-        start_time: Optional[int] = None,
-        record_exception: bool = True,
-        set_status_on_exception: bool = True,
-        *,
-        openinference_span_kind: Optional[OpenInferenceSpanKind] = None,
-    ) -> OpenInferenceSpan:
-        span: Span
-        if get_value(_SUPPRESS_INSTRUMENTATION_KEY):
-            span = INVALID_SPAN
-        else:
-            tracer = cast(Tracer, self.__wrapped__)
-            span = tracer.__class__.start_span(
-                self,
-                name=name,
-                context=context,
-                kind=kind,
-                attributes=None,
-                links=links,
-                start_time=start_time,
-                record_exception=record_exception,
-                set_status_on_exception=set_status_on_exception,
-            )
-        span = OpenInferenceSpan(span, config=self._self_config)
-        if attributes:
-            span.set_attributes(dict(attributes))
-        if openinference_span_kind is not None:
-            span.set_attributes(get_span_kind_attributes(openinference_span_kind))
-        span.set_attributes(dict(get_attributes_from_context()))
-        return span
-
-    @overload  # for @tracer.agent usage (no parameters)
-    def agent(
-        self,
-        wrapped_function: Callable[ParametersType, ReturnType],
-        /,
-        *,
-        name: None = None,
-    ) -> Callable[ParametersType, ReturnType]: ...
-
-    @overload  # for @tracer.agent(name="name") usage (with parameters)
-    def agent(
-        self,
-        wrapped_function: None = None,
-        /,
-        *,
-        name: Optional[str] = None,
-    ) -> Callable[[Callable[ParametersType, ReturnType]], Callable[ParametersType, ReturnType]]: ...
-
-    def agent(
-        self,
-        wrapped_function: Optional[Callable[ParametersType, ReturnType]] = None,
-        /,
-        *,
-        name: Optional[str] = None,
-    ) -> Union[
-        Callable[ParametersType, ReturnType],
-        Callable[[Callable[ParametersType, ReturnType]], Callable[ParametersType, ReturnType]],
-    ]:
-        return self._chain(
-            wrapped_function,
-            kind=OpenInferenceSpanKindValues.AGENT,  # chains and agents differ only in span kind
-            name=name,
-        )
-
-    @overload  # for @tracer.chain usage (no parameters)
-    def chain(
-        self,
-        wrapped_function: Callable[ParametersType, ReturnType],
-        /,
-        *,
-        name: None = None,
-    ) -> Callable[ParametersType, ReturnType]: ...
-
-    @overload  # for @tracer.chain(name="name") usage (with parameters)
-    def chain(
-        self,
-        wrapped_function: None = None,
-        /,
-        *,
-        name: Optional[str] = None,
-    ) -> Callable[[Callable[ParametersType, ReturnType]], Callable[ParametersType, ReturnType]]: ...
-
-    def chain(
-        self,
-        wrapped_function: Optional[Callable[ParametersType, ReturnType]] = None,
-        /,
-        *,
-        name: Optional[str] = None,
-    ) -> Union[
-        Callable[ParametersType, ReturnType],
-        Callable[[Callable[ParametersType, ReturnType]], Callable[ParametersType, ReturnType]],
-    ]:
-        return self._chain(wrapped_function, kind=OpenInferenceSpanKindValues.CHAIN, name=name)
-
-    def _chain(
-        self,
-        wrapped_function: Optional[Callable[ParametersType, ReturnType]] = None,
-        /,
-        *,
-        kind: OpenInferenceSpanKindValues,
-        name: Optional[str] = None,
-    ) -> Union[
-        Callable[ParametersType, ReturnType],
-        Callable[[Callable[ParametersType, ReturnType]], Callable[ParametersType, ReturnType]],
-    ]:
-        @wrapt.decorator  # type: ignore[misc]
-        def sync_wrapper(
-            wrapped: Callable[ParametersType, ReturnType],
-            instance: Any,
-            args: Tuple[Any, ...],
-            kwargs: Dict[str, Any],
-        ) -> ReturnType:
-            tracer = self
-            with _chain_context(
-                tracer=tracer,
-                name=name,
-                kind=kind,
-                wrapped=wrapped,
-                instance=instance,
-                args=args,
-                kwargs=kwargs,
-            ) as chain_context:
-                output = wrapped(*args, **kwargs)
-                return chain_context.process_output(output)
-
-        @wrapt.decorator  #  type: ignore[misc]
-        async def async_wrapper(
-            wrapped: Callable[ParametersType, Awaitable[ReturnType]],
-            instance: Any,
-            args: Tuple[Any, ...],
-            kwargs: Dict[str, Any],
-        ) -> ReturnType:
-            tracer = self
-            with _chain_context(
-                tracer=tracer,
-                name=name,
-                kind=kind,
-                wrapped=wrapped,
-                instance=instance,
-                args=args,
-                kwargs=kwargs,
-            ) as chain_context:
-                output = await wrapped(*args, **kwargs)
-                return chain_context.process_output(output)
-
-        if wrapped_function is not None:
-            if asyncio.iscoroutinefunction(wrapped_function):
-                return async_wrapper(wrapped_function)  # type: ignore[no-any-return]
-            return sync_wrapper(wrapped_function)  # type: ignore[no-any-return]
-        return lambda f: async_wrapper(f) if asyncio.iscoroutinefunction(f) else sync_wrapper(f)
-
-    @overload  # for @tracer.tool usage (no parameters)
-    def tool(
-        self,
-        wrapped_function: Callable[ParametersType, ReturnType],
-        /,
-        *,
-        name: None = None,
-        description: Optional[str] = None,
-        parameters: Optional[Union[str, Dict[str, Any]]] = None,
-    ) -> Callable[ParametersType, ReturnType]: ...
-
-    @overload  # for @tracer.tool(name="name") usage (with parameters)
-    def tool(
-        self,
-        wrapped_function: None = None,
-        /,
-        *,
-        name: Optional[str] = None,
-        description: Optional[str] = None,
-        parameters: Optional[Union[str, Dict[str, Any]]] = None,
-    ) -> Callable[[Callable[ParametersType, ReturnType]], Callable[ParametersType, ReturnType]]: ...
-
-    def tool(
-        self,
-        wrapped_function: Optional[Callable[ParametersType, ReturnType]] = None,
-        /,
-        *,
-        name: Optional[str] = None,
-        description: Optional[str] = None,
-        parameters: Optional[Union[str, Dict[str, Any]]] = None,
-    ) -> Union[
-        Callable[ParametersType, ReturnType],
-        Callable[[Callable[ParametersType, ReturnType]], Callable[ParametersType, ReturnType]],
-    ]:
-        @wrapt.decorator  # type: ignore[misc]
-        def sync_wrapper(
-            wrapped: Callable[ParametersType, ReturnType],
-            instance: Any,
-            args: Tuple[Any, ...],
-            kwargs: Dict[str, Any],
-        ) -> ReturnType:
-            tracer = self
-            with _tool_context(
-                tracer=tracer,
-                name=name,
-                description=description,
-                parameters=parameters,
-                wrapped=wrapped,
-                instance=instance,
-                args=args,
-                kwargs=kwargs,
-            ) as tool_context:
-                output = wrapped(*args, **kwargs)
-                return tool_context.process_output(output)
-
-        @wrapt.decorator  #  type: ignore[misc]
-        async def async_wrapper(
-            wrapped: Callable[ParametersType, Awaitable[ReturnType]],
-            instance: Any,
-            args: Tuple[Any, ...],
-            kwargs: Dict[str, Any],
-        ) -> ReturnType:
-            tracer = self
-            with _tool_context(
-                tracer=tracer,
-                name=name,
-                description=description,
-                parameters=parameters,
-                wrapped=wrapped,
-                instance=instance,
-                args=args,
-                kwargs=kwargs,
-            ) as tool_context:
-                output = await wrapped(*args, **kwargs)
-                return tool_context.process_output(output)
-
-        if wrapped_function is not None:
-            if asyncio.iscoroutinefunction(wrapped_function):
-                return async_wrapper(wrapped_function)  # type: ignore[no-any-return]
-            return sync_wrapper(wrapped_function)  # type: ignore[no-any-return]
-        return lambda f: async_wrapper(f) if asyncio.iscoroutinefunction(f) else sync_wrapper(f)
-
-
-class TracerProvider(OTelTracerProvider):
-    def __init__(
-        self,
-        *args: Any,
-        config: Optional[TraceConfig] = None,
-        **kwargs: Any,
-    ) -> None:
-        super().__init__(*args, **kwargs)
-        self._oi_trace_config = config or TraceConfig()
-
-    def get_tracer(
-        self,
-        *args: Any,
-        **kwargs: Any,
-    ) -> OITracer:
-        tracer = super().get_tracer(*args, **kwargs)
-        return OITracer(tracer, config=self._oi_trace_config)
 
 
 def is_base64_url(url: str) -> bool:
@@ -949,7 +565,7 @@ def _is_dataclass_instance(obj: Any) -> TypeGuard["DataclassInstance"]:
 
 
 class _ChainContext:
-    def __init__(self, span: OpenInferenceSpan) -> None:
+    def __init__(self, span: "OpenInferenceSpan") -> None:
         self._span = span
 
     def process_output(self, output: ReturnType) -> ReturnType:
@@ -993,7 +609,7 @@ def _chain_context(
 
 
 class _ToolContext:
-    def __init__(self, span: OpenInferenceSpan) -> None:
+    def __init__(self, span: "OpenInferenceSpan") -> None:
         self._span = span
 
     def process_output(self, output: ReturnType) -> ReturnType:
